requirements.txt
django==3.0.8
django-python3-ldap==0.11.3
djangorestframework==3.11.0
django-cors-headers==3.4.0
psycopg2==2.8.6